+++ src/index.js
@@ -1,32 +1,19 @@
 import Swal from 'sweetalert2';
 import createSpaceships from './modules/spaceships.js';
 import showNotification from './modules/notification.js';
+import mobileMenu from './modules/mobileMenu.js';
+import insertLogo from './modules/insertLogo.js';
 import { getLikes, likeItem } from './modules/apiInvolvement.js';
 import './style.css';
 import createCard from './modules/card.js';
 import fetchMovies from './modules/apiTvmaze.js';
-import logo from './images/logo2.png';
 import commentsPopup from './modules/commentsPopup.js';
 import cardCounter from './modules/cardCounter.js';
 
 createSpaceships();
 showNotification();
-
-document.getElementById('menu-toggle').addEventListener('click', function () {
-  const nav = document.getElementById('nav');
-  const navigationLeft = getComputedStyle(nav).getPropertyValue('left');
-  if (navigationLeft === '0px') {
-    nav.style.left = '-100%';
-  } else {
-    nav.style.left = '0px';
-  }
-});
-
-const logoContainer = document.querySelector('.logo');
-const imgElement = document.createElement('img');
-imgElement.src = logo;
-
-logoContainer.appendChild(imgElement);
+insertLogo();
+mobileMenu();
 
 const movieContainer = document.querySelector('.movies');
